+++ Source/SwiftLintFramework/Rules/Lint/WeakDelegateRule.swift
@@ -26,7 +26,8 @@ public struct WeakDelegateRule: OptInRule, ASTRule, ConfigurationProviderRule {
             Example("class Foo {\n protocol P {\n var delegate: AnyObject? { get set }\n}\n}\n"),
             Example("class Foo {\n var computedDelegate: ComputedDelegate {\n return bar() \n} \n}"),
             Example("struct Foo {\n @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate \n}"),
-            Example("struct Foo {\n @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate \n}")
+            Example("struct Foo {\n @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate \n}"),
+            Example("struct Foo {\n @WKExtensionDelegateAdaptor(ExtensionDelegate.self) var extensionDelegate \n}")
         ],
         triggeringExamples: [
             Example("class Foo {\n  ↓var delegate: SomeProtocol?\n}\n"),
@@ -75,7 +76,9 @@ public struct WeakDelegateRule: OptInRule, ASTRule, ConfigurationProviderRule {
                 let offset = attribute.offset,
                 let length = attribute.length,
                 let value = file.stringView.substringWithByteRange(ByteRange(location: offset, length: length)),
-                (value.hasPrefix("@UIApplicationDelegateAdaptor") || value.hasPrefix("@NSApplicationDelegateAdaptor")) {
+                (value.hasPrefix("@UIApplicationDelegateAdaptor")
+                  || value.hasPrefix("@NSApplicationDelegateAdaptor")
+                  || value.hasPrefix("@WKExtensionDelegateAdaptor")) {
                 return []
             }
         }